backend/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1